added error message display after unsuccessfull ticket deletion

+++ app/views.py
@@ -222,6 +222,8 @@ class SearchSeancesView(generic.ListView):
 
 
 def delete_ticket(request):
+    global error
+    error = ''
     if request.method == 'POST':
         form = forms.DeleteTicketForm(request.POST)
 
@@ -229,18 +231,21 @@ def delete_ticket(request):
             ticket_number = form.cleaned_data['ticket_number']
             phone_number = form.cleaned_data['phone_number']
 
-            ticket = Tickets.objects.get(pk=ticket_number)
+            try:
+                ticket = Tickets.objects.get(pk=ticket_number)
+            except Tickets.DoesNotExist:
+                ticket = None
 
-            # correct data, so you can delete ticket
+            # correct data - you can delete ticket
             # otherwise show some message
-            # TODO: pokaz wiadomosc ze nie udalo sie usunac biletu
-            if ticket.client.phone_number == phone_number:
+            if ticket is not None and ticket.client.phone_number == phone_number:
                 ticket.delete()
             else:
                 error = 'Nie udało się usunąc biletu, niepoprawne dane'
+                return render(request, 'app/cancel_ticket.html', {'form': form, 'error': error})
 
             return redirect('index')
     else:
         form = forms.DeleteTicketForm()
 
-    return render(request, 'app/cancel_ticket.html', {'form': form})
+    return render(request, 'app/cancel_ticket.html', {'form': form, 'error': error})
